refactor: split decision emit invariant target
